dom/xml/test/test_bug392338.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=392338
-->
<head>
  <title>Test for Bug 392338</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=392338">Mozilla Bug 392338</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 392338 **/

SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  var doc = document.implementation.createDocument("", "", null);

  var chromeScriptUrl = SimpleTest.getTestFileURL("file_bug392338.js");
  var script = SpecialPowers.loadChromeScript(chromeScriptUrl);

  script.addMessageListener('modify-request-completed', function ml(data) {
    ok(data, "XMLDocument should have notificationCallbacks");
    script.removeMessageListener('modify-request-completed', ml);
    SimpleTest.finish();
  });

  doc.load(window.location.href);
});

</script>
</pre>
</body>
</html>
